# backend/generated/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "",
    "description": "The main API structure\n\n- `/api/v1beta/` <- Most of the API is nested under here. All of the resources there are scoped to what is accessible by the authenticated identity.\n- `/api/v1beta/me` <- Everything under this path is scoped to the subject of the authenticated identity.\nThis means that the identity may be authorized to view more resources, but this is the default view for them.\nE.g. the chats route scoped under there will only list the chats created by the user, but the user may be authorized to also view chats shared by other users.\n",
    "version": ""
  },
  "paths": {
...
    "/api/v1beta/me/profile": {
      "get": {
        "tags": [],
        "operationId": "profile",
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserProfile"
                }
              }
            }
          },
          "401": {
            "description": "When no valid JWT token is provided"
          }
        },
        "security": [
          {
            "bearer_auth": []
          }
        ]
      }
    },
...
    "/api/v1beta/share-grants": {
      "get": {
        "tags": [
          "share_grants"
        ],
        "summary": "List share grants for a resource",
        "operationId": "list_share_grants",
        "parameters": [
          {
            "name": "resource_type",
            "in": "query",
            "description": "The type of resource (e.g., 'assistant')",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "resource_id",
            "in": "query",
            "description": "The ID of the resource",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successfully retrieved share grants",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ListShareGrantsResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid query parameters"
          },
          "401": {
            "description": "When no valid JWT token is provided"
          },
          "403": {
            "description": "User does not own the resource"
          },
          "500": {
            "description": "Server error"
          }
        },
        "security": [
          {
            "bearer_auth": []
          }
        ]
      },
      "post": {
        "tags": [
          "share_grants"
        ],
        "summary": "Create a new share grant",
        "operationId": "create_share_grant",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateShareGrantRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successfully created the share grant",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CreateShareGrantResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request data"
          },
          "401": {
            "description": "When no valid JWT token is provided"
          },
          "403": {
            "description": "User does not own the resource"
          },
          "500": {
            "description": "Server error"
          }
        },
        "security": [
          {
            "bearer_auth": []
          }
        ]
      }
    },
    "/api/v1beta/share-grants/{grant_id}": {
      "delete": {
        "tags": [
          "share_grants"
        ],
        "summary": "Delete a share grant",
        "operationId": "delete_share_grant",
        "parameters": [
          {
            "name": "grant_id",
            "in": "path",
            "description": "The ID of the share grant to delete",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successfully deleted the share grant"
          },
          "400": {
            "description": "Invalid grant ID format"
          },
          "401": {
            "description": "When no valid JWT token is provided"
          },
          "403": {
            "description": "User does not own the resource"
          },
          "404": {
            "description": "Share grant not found"
          },
          "500": {
            "description": "Server error"
          }
        },
        "security": [
          {
            "bearer_auth": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CreateShareGrantRequest": {
        "type": "object",
        "description": "Request to create a new share grant",
        "required": [
          "resource_type",
          "resource_id",
          "subject_type",
          "subject_id_type",
          "subject_id",
          "role"
        ],
        "properties": {
          "resource_id": {
            "type": "string",
            "description": "The ID of the resource to share"
          },
          "resource_type": {
            "type": "string",
            "description": "The type of resource to share (e.g., \"assistant\")"
          },
          "role": {
            "type": "string",
            "description": "The role to grant (e.g., \"viewer\")"
          },
          "subject_id": {
            "type": "string",
            "description": "The ID of the subject to grant access to"
          },
          "subject_id_type": {
            "type": "string",
            "description": "The type of subject ID (e.g., \"id\" or \"oidc_issuer_and_subject\")"
          },
          "subject_type": {
            "type": "string",
            "description": "The type of subject to grant access to (e.g., \"user\")"
          }
        }
      },
      "CreateShareGrantResponse": {
        "allOf": [
          {
            "$ref": "#/components/schemas/ShareGrant",
            "description": "The created share grant"
          }
        ],
        "description": "Response when creating a share grant"
      },
...
      "ListShareGrantsResponse": {
        "type": "object",
        "description": "Response when listing share grants",
        "required": [
          "grants"
        ],
        "properties": {
          "grants": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ShareGrant"
            },
            "description": "The list of share grants"
          }
        }
      },
...
      "ShareGrant": {
        "type": "object",
        "description": "A share grant model",
        "required": [
          "id",
          "resource_type",
          "resource_id",
          "subject_type",
          "subject_id_type",
          "subject_id",
          "role",
          "created_at",
          "updated_at"
        ],
        "properties": {
          "created_at": {
            "type": "string",
            "format": "date-time",
            "description": "When this share grant was created"
          },
          "id": {
            "type": "string",
            "description": "The unique ID of the share grant"
          },
          "resource_id": {
            "type": "string",
            "description": "The ID of the resource being shared"
          },
          "resource_type": {
            "type": "string",
            "description": "The type of resource being shared (e.g., \"assistant\")"
          },
          "role": {
            "type": "string",
            "description": "The role being granted (e.g., \"viewer\")"
          },
          "subject_id": {
            "type": "string",
            "description": "The ID of the subject being granted access"
          },
          "subject_id_type": {
            "type": "string",
            "description": "The type of subject ID (e.g., \"id\" or \"oidc_issuer_and_subject\")"
          },
          "subject_type": {
            "type": "string",
            "description": "The type of subject being granted access (e.g., \"user\")"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "description": "When this share grant was last updated"
          }
        }
      },
...
      "UserProfile": {
        "type": "object",
        "required": [
          "id",
          "preferred_language",
          "groups",
          "organization_group_ids"
        ],
        "properties": {
          "email": {
            "type": "string",
            "description": "The user's email address. Shouldn't be used as a unique identifier, as it may change."
          },
          "groups": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "description": "List of groups the user belongs to.\n\nThis is derived from the `groups` claim in the ID token.\nIf the claim is not present, this will be an empty list."
          },
          "id": {
            "type": "string"
          },
          "name": {
            "type": "string",
            "description": "The user's display name."
          },
          "organization_group_ids": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "description": "Organization group IDs from the `groups` claim.\n\nThese can be used as subject_id when creating share grants\nwith subject_id_type \"organization_group_id\"."
          },
          "organization_user_id": {
            "type": "string",
            "description": "Organization user ID from the `oid` claim (Entra ID specific).\n\nThis can be used as the subject_id when creating share grants\nwith subject_id_type \"organization_user_id\"."
          },
          "picture": {
            "type": "string",
            "description": "The user's profile picture URL."
          },
          "preferred_language": {
            "type": "string",
            "description": "The user's preferred language.\n\nThe final determined language is intersected with our supported languages, to determine the final language.\n\nWill be a BCP 47 language tag (e.g. \"en\" or \"en-US\").\n\nThis is derived in the following order (highest priority first):\n- ID token claims\n- Browser Accept-Language header\n- Default to \"en\""
          }
        }
      },
...
    }
  }
}